campus/requirements.txt
Django
django-environ

# AI / LLM
langchain
langchain-google-genai
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
proto-plus
protobuf

# Vector / NLP
faiss-cpu
numpy
tqdm

# Documents & OCR
PyPDF2
pdf2image
PyMuPDF
pillow
pytesseract
python-docx
reportlab

# Web / HTTP
requests
httpx

# Payments
razorpay

# DB / Django extras
psycopg2-binary
sqlparse
tzdata

# Utilities
python-dotenv
typing_extensions
typing-inspection
packaging
six
